Remove flag for logfile and always log to file
The flag to disable/enable file logging was removed, because a windows
application can not log to stdout on windows. Therefore it is necessary
to always log to a logfile in order to retrieve any logs. This is the
reason the logfile flag was removed and it is always logged to file.