Fix broken scrolling
The defaultusername browser introduced an issue with scrolling inside
the application. Not the actual widget was hidden, but the "internal"
widget was hidden and the scroll window still visible. This introduced
an issue for every other scroll widget, because it stole the "scroll
focus". Therefore the actual widget is now hidden.